backend/requirements.txt
--index-url https://download.pytorch.org/whl/cpufastapi==0.104.1

fastapi==0.104.1uvicorn[standard]==0.24.0

uvicorn[standard]==0.24.0python-dotenv==1.0.0

python-dotenv==1.0.0torch==2.1.0

torch==2.1.0torchvision==0.16.0

torchvision==0.16.0Pillow==10.1.0

Pillow==10.1.0numpy==1.24.3

numpy==1.24.3pydantic==2.5.0

pydantic==2.5.0opencv-python-headless==4.8.1.78

opencv-python-headless==4.8.1.78